requirements.txt
requests==2.31.0
google-generativeai==0.3.2
flask==2.3.3
python-dateutil==2.8.2